+++ src/mln/plugin/plugin_shader.cpp
@@ -93,7 +93,8 @@ public:
 
     gfx::ShaderPtr getOrCreateShader(gfx::Context& context,
                                      const StringIDSetsPair& propertiesAsUniforms,
-                                     std::string_view) override {
+                                     std::string_view,
+                                     gfx::ProjectionVariant) override {
         const auto name = getShaderName(groupName, propertyHash(propertiesAsUniforms));
         if (auto existing = getShader(name)) return existing;
         const auto pluginPrelude = resourcePrelude(*definition) + propertyPrelude(*definition, propertiesAsUniforms);